refactor: inline StreamMessageValidator
